Bug#30151330 - SER_BUFF SHOULD NOT HAVE ITS TOP VALUE NO LESS THAN ITS END VALUE Bug#30625357 - EXPLAIN CORRECT REASON IN TRACE WHEN COLUMN IS NOT LOADED TO SECONDARY ENGINE Bug#30650153 - QUERY WITH MULTIPLE JOINS DOES NOT OFFLOAD WHEN PLAN BECOMES TOO LARGE
Bug 30151330, Bug 30650153:
The queries are not offloaded as the query execution plan buffer space is exhausted.
Added proper log messages to specify space availability/requirement units.
Bug 30625357:
Though error buffer is updated with the error "column marked NOT SECONDARY",
for explain queries, errors are not posted on the client side as query could be
re-planned without secondary engine. As a fix, added trace messages also.

Change-Id: Ic52faf072cb82a6fd033bcd25d28d4747a4f2140