Bug #30293495 NDB : EXCESSIVE LOGGING ABOUT NODEID ALLOCATION FAILURES
A nodeID allocation request can fail with temporary errors such as
1) NotMaster: the president node has not yet become the master
and cannot allocate a node ID
2) Busy: the master node is busy with another request
3) NodeFailureHandlingNotCompleted: a node restart is in progress

For 'Busy' and 'NodeFailureHandlingNotCompleted' errors, the data
node always sleeps 100ms before retrying the nodeID alloc. But
on receiving a 'NotMaster' error, the nodeID alloc is always
retried immediately. This causes a very high volume of retries per
second in the 'NotMaster' case, which can be observed as a high
volume of 'Alloc node id for node XYZ failed" log messages, e.g.
15000 logs per second.

This is fixed by setting the retry behaviour as per the cause of
the 'NotMaster' error
1) Master not yet elected
Get nodeID of another node and retry immediately
2) Master elected but master node not ready due to DIH
Always sleep 1 second between retries.
3) Master elected, but queried node is not master
Get nodeID of master node and retry immediately

Approved-by: Frazer Clement <frazer.clement@oracle.com>

Change-Id: I4b73e9a65b9fe2f03b2e25a9c719f71481bed43c