Bug #20644698 INNODB TABLESPACE IMPORT FAILS WHEN IMPORTING TABLE W/ DIFFERENT DATA DIRECTORY
Problem:
--------
During tablespace import, If the source table was created  by
specifying the DATA DIRECTORY clause, then the table that we replace
on the destination instance must be defined with the same DATA
DIRECTORY clause. If the clauses do not match then schema mismatch
error is reported.
Issue-1: The error message do not specify the exact reason of this
error.
Issue-2: If .cfg file is not found during import then we do not
validate table flags to fsp flags. Later due to table flags
mismatch, server terminates due to fatal error.

Solution:
---------
Fix-1: If .cfg file is available during import then check if  the
source table meta-data (.cfg file) flag for data_dir do not match
with server table flag for data_dir then throw an error and
terminate import operation.

Fix-2: If .cfg file is not available during import then check if the
source table fsp (.ibd file) flag  for data_dir do not match with
server table flag for data_dir then throw an error and terminate
import operation.

Note:
-----
This fix is a backport into 5.6 for a fix (of #30190199 & #30190227 bugs)
in 5.7 and above.

RB: 23532
Reviewed by : Kevin Lewis <kevin.lewis@oracle.com>