Bug#30600320 INVALID YEAR VALUE EXTRACTED FROM A DATE COLUMN
When a Date value is read back from NDB, ClusterJ sometimes extracts
wrong year values from the row. The Utility class, when unpacking the
Date value, wrongly extracts a few more bits in addition to the bits
storing the year and thus causing this issue. Fixed that by extracting
only the required bits. Also updated a few testcases to verify the fix.

Change-Id: I5e388eeb83305761e732f9b0c7d23f6d03026070