Bug #30401416 RWLOCK:REFINE LOCK->RECURSIVE WITH C11 ATOMICS
- In rw_lock_t code we ensure ordering of access to the
boolean recursive flag and the writer thread id by using
a combination of gcc builtins and os_mutex when the
built ins are not avaialable. This is modern c++ so this
patch changes some of the places that are trivial to
convert to using std::atomic.
- This is a contribution from Yibo Cai from ARM, however
the original commit was modified, the proposed os_atomic
class was removed in code review. The implementation
provided broke our debug tests due to a faulty operator=
implementation.
- This operator was required due to a dubious copy of
the rw_lock_t in gis0sea.cc. This copy was removed as
the rw_lock_t does not have well defined copy semantics,
similarly to std::mutex.

RB#23462 Approved by Jakub Lopuszanski <jakub.lopuszanski@oracle.com>