WL#13727 REMOVES NDB_WAITER DEPENDENCY ON NDBT
Removes dependencies on NDBT library which is meant to be used for
testing only.

ndb_waiter relies on NDBT library for returning status codes and
printing 'NDBT_ProgramExit: Code - status'. It also uses methods like
g_info and g_err to print messages which are defined in the same
library.

Back-porting NdbToolsProgramExitCodes.hpp which was introduced in
WL#13223 to versions 8.0 and above that defines common error codes
to replace those defined in NDBT library. Replacing g_info and
g_err with ndbout and ndberr respectively.

Change-Id: I49098549aac17355558068a6ed5b2ce047c0072b