Bug#30691682 MIN/MAX ISSUE WITH NOT EXISTS SUBQUERY
The bug's query is changed to
SELECT MAX(t1.a) FROM t1 ANTIJOIN t2 ON t1.a=t2.a;
optimize_aggregated_query() has a test on LEFT JOIN: if the inner
table is referenced by WHERE it bails out.
An antijoin is implemented as a kind of left join, but the WHERE is empty
(even though we ultimately, later in the optimizer stage, add
a triggered condition "t2.a IS NULL", it's not present at this stage).
So, this function proceeds, calculates aggregates and thus returns a wrong
result (and EXPLAIN says "Select tables optimized away").

Fix: this function is designed to optimize this:
SELECT COUNT(*) FROM join-of-tables no-WHERE;
or
SELECT MAX(x) FROM join-of-tables WHERE indexed_column=constant;
and I cannot possibly see how this could successfully apply to
a query with antijoin or semijoin; consider that it doesn't even
apply to this simplified query with inner join:
SELECT max(n) FROM numbers, ids WHERE ids.id = numbers.n;
due to equality between columns.

So, I simply disable the optimization if we have semijoin or antijoin.

Change-Id: I311748328c70c794f759117bc1d9b14b8870b18e