Bug #30694177 RW_LOCK_X_LOCK_LOW: CONDITIONAL JUMP OR MOVE...
Valgrind complains that unitialized value of lock->writer_thread is used in conditional jump, during evaluation of:
```
   if (recursive && os_thread_eq(writer_thread, thread_id))
```

We (correctly) maintain the invariant that if `lock->recursive` is true then `lock->writer_thread` must be already initialized.
However, `recursive` and `writer_thread` are local variables with copies of the values of `lock->recursive` and `lock->writer_thread` respectively,
and since they are on stack, the compiler can reorder LHS and RHS of the && operator using the "as-if" rule, because
there should be no observable effect of accessing on-stack variables in different order.
This patch changes `writer_thread` to direct read from `lock->writer_thread` which forces the compiler to generate assembly
in which RHS is only read if the guarding LHS expression evaluates to true.

RB:23651
Reviewed-by: Nikša Skeledžija <niksa.skeledzija@oracle.com>