Bug#30503401 NDB_DDL_OPEN_TRANS AND NDB_RPL_DDL_OPEN_TRANS TESTS FAIL SPORADICALLY
Problem:
--------

The ndb_ddl_open_trans and ndb_rpl_ddl_open_trans tests fail sporadically
with the following warning appearing in the mysqld error log:

1062: Duplicate entry 'ndbcluster-13' for key 'tables.engine'

The above tests provoke failures in schema distribution of DDL changes.
As a result, these anomalous situations are detected and addressed by
the auto sync mechanism. More specifically, synchronization of a table
after the failed distribution of an ALTER RENAME statement leads to the
above warning being spewed out.

Fix:
----

The warning is harmless and auto sync does not cause any change in
behaviour. The auto sync mechanism could be turned off for the tests
but given that there's no change in behaviour, it is prudent to simply
suppress the warning instead.

Change-Id: I9ed8fb43ae203267383d8942ab6f92a594fcacaf