WL#13728 REMOVES NDB_SHOW_TABLES DEPENDENCY ON NDBT
Removes the dependency of ndb_show_tables tool on NDBT library which is to
be used for testing only.

NDBT library was used by ndb_show_tables to return status codes and
print them using 'NDBT_ProgramExit'. The status codes were printed for
testing purposes and it is not necessary to print out the status codes.

In the current change "--echo" is added in the test files to
maintain a new line between executions of ndb_show_tables.

Change-Id: Ie1318351191f729e4f0445d149825a1deacc5762