Bug#30613385: NDB_BINLOG_DISCOVER TEST FAILS SPORADICALLY IN PB2
The test fails sporadically due to the following error message in
the mysql server log:

[ERROR] [MY-010865] [Server] NDB: Failed to setup events for
'mysql.ndb_schema' table

This is due to the test not waiting properly before both the mysql
servers used in the test are able to reconnect with the datanodes after
a system restart. This patch fixes that and refactors the test to
improve readability.

Change-Id: I6aa7a416c540a99959155bf7a1566f2dabbf3f5d