WL#7636: Fix a bug found in late RQG testing:
We didn't correctly include tables from join_nest(s), not being part of
our set of upper_nests, in the set of 'embedded_nest()' which we
are allowed to refer tables from.

This caused that in rare cases we failed to recognice a table
as not pushable.

Fixing this also allowed us to remove the 'm_outer_nest' table map, as
such 'outer' tables are now included in the upper_nest table map instead.

Reviewed by: Mikael Ronstrom <mikael.ronstrom@oracle.com>