Bug#30737242 do not link unit-tests against http plugins
Problem
-------

unittests and the rest_api_lib depend on functionality provided by the
http_backend plugin.

On windows the DLL search path looks for executables in the exectutables
directory, PATH, ...  while on Unix the search path can be easily
embedded into the executable.

To the allow running the unittests easily the http-backend plugin was
copied into the runtime_output_directory/ on windows only.

This approach leads to problems as soon as a DLL is loaded as plugin and
as library and static/globals symbols are provided by the library:

- runtime_output_directory/http_backend.dll
- plugin_output_directory/http_backend.dll

LoadLibrary() would only load the plugin/library once give they refer to
the same file, but as they files are in different directories,
LoadLibrary() will load them twice and load to different exectuables
seeing different addresses for the globals.

Change
------

- split library part out of the http_auth_realm plugin
- link rest-api-lib against http_auth_realm_lib instead of the
  http_server plugin

RB: 23537
Reviewed by: Andrzej Religa <andrzej.religa@oracle.com>