Bug#30712243 - MEMORY LEAK FROM CREATE_HANDLER WITH MATERIALIZED                TEMPORARY TABLES?
Problem:
While removing a where condition because one of the participating condition
is always false, a materialized derived table does not get cleaned
up resulting in a memory leak.

Analysis:
In clean_up_after_removal, before deleting a subquery, current query block
is unlinked from the global selects (function "exclude tree"). While
unlinking, all the inner query expressions of the current block are
unlinked before the cleanup of the current query block is performed. But
the unlinked inner query expression is not cleaned up.
For the query in bugpage, when function "exclude_tree" is called for the
outermost select in the "exists" clause, it unlinks the first derived table.
As a result, cleanup does not happen for this derived table.

Fix:
Cleanup inner query expressions after unlinking them from the query block.
So we replace function "exclude_level" with "exclude_tree" which does both
unlinking and cleanup.

Reviewed by : Roy Lyseng <roy.lyseng@oracle.com>