BUG#30259491 REFACTOR PROCESSES FUNCTIONALITY TO CLASS
Refactor functionality related with cluster processes to a new class,
ProcessManagement. This patch only includes refactoring and some minor
changes needed by the object-oriented approach. Changes to the
functionality will be addressed in following patches.

The class ProcessManagement was created to gather functionality related
with managing cluster processes which was previously in 'main.cpp'. This
is the first step of several changes which will allows us to provide a
cleaner API and manage cluster processes status inside the class.
Methods used by other modules were made public while the rest remained
private. It is expected that further refactoring reduces the number of
public methods as more functionality is encapsulated inside the class.
Many methods depended on 'g_config' and 'g_setup_progname' which were
global variables of 'main.cpp'. Those are mapped to instance variables
of ProcessManagement and instantiated in the class constructor. In the
case of 'g_config' we keep a reference to the external configuration so
changes from outside the class can be visible inside and vice-versa.
This is a bad object-oriented practice but necessary until the
configuration itself is broken down into classes.

Some other changes were introduced:
- Some constants were moved from 'main.cpp' to ProcessManagement because
they are mainly used as parameters of the refactored methods.
- Method names remain the same, only changing to camel case convention.
- A variable of type  ProcessManagement was created in the main function
to handle operations with processes. This variable had to be passed as
parameter to several functions, which prototype became dependent of
ProcessManagement. In order to avoid circular imports of header files,
their declaration was moved from 'atrt.hpp' to 'main.cpp'.

The code division in this patch is sometimes a little bit arbitrary
since the code is not truly object-oriented. The data is still hold in
the existing data structures which makes decoupling difficult. Only
after breaking those data structures and migrate them into classes
better code is possible.

Change-Id: If41eea2cb9e46807348efb975c73fc47938d6153