Bug#30776375 - SYMBOLS DEFINED TWICE WITH SOME OF THE P_S INSTRUMENTATION HEADERS
Before this fix,
compiling a component using performance schema services failed
depending on the platform,
for non trivial (more than one *.cc file) components.

The root cause is the incorrect declaration for service placeholders.

For example, in include/mysql/components/services/mysql_mutex.h,
the code used:
  REQUIRES_SERVICE_PLACEHOLDER(mysql_mutex_v1);

This is incorrect: it is a definition, not a declaration,
as the extern keyword is missing.

With this fix,

1)

The doxygen documentation for REQUIRES_SERVICE_PLACEHOLDER()
is reworded, for clarity.

2)

In the header files for performance schema services,
the declaration of service placeholders is changed to:

  REQUIRES_SERVICE_PLACEHOLDER(xxx_vn)

  extern REQUIRES_XXX_SERVICE_PLACEHOLDER;

For example,
the macro REQUIRES_MYSQL_MUTEX_SERVICE_PLACEHOLDER is provided,
so a component can define the holder, without hard coding
the service exact name (like mysql_mutex_v1).

To help components to declare service dependencies,
the following macro is also provided (see note (a)):

  REQUIRES_SERVICE(xxx_vn)

For example,
the macro REQUIRES_PSI_MEMORY_SERVICE
can be used to to declare the dependency
on the performance schema memory service,
without hard coding the service exact name (like psi_memory_v1).

Note (a)
This macro was sometime already provided,
but not consistently.
With this fix, every service comes with the appropriate macro(s).

3)

The component implementation for:
- pfs_example
- validate_password
was building by accident,
as definitions for required place holders were missing.

With the fix, the component definition is complete.

4)

Last, some code cleanup is done in helpers for
- mysql_mutex
- mysql_cond
- mysql_rwlock
to avoid repeating the full service name (mysql_service_mysql_mutex_v1)
all the time, and use a helper like MYSQL_MUTEX_CALL() instead.

This has no impact on user code, and prevents merge conflicts,
should the service version change.

Approved by: Georgi Kodinov <georgi.kodinov@oracle.com>