Bug #30787515: SUPPORT ABORTING EXPLAIN ANALYZE
Some queries, in particular misoptimized ones, can take a very long time to
execute. For those, it is very useful to start EXPLAIN ANALYZE, let it run for
a little while, Ctrl-C (KILL QUERY) it and then get out an idea of where the
time was spent so far. This will pinpoint many kinds of misestimations and the
likes without having the run it to completion.

Also tweak the client slightly so that it does not suppress the single row that
the server will output in this case (it already has fetched the row and has
printed the header, so printing one row will not make the Ctrl-C return appreciably
slower).

Change-Id: I8f1aaeb7ee39601b76b87a8bce89d13b5cf94b35