Bug#30753524 DBUG_ASSERT((MAP_OF_NEW_FDS & PSEUDO_TABLE_BITS) == 0)
Now that MySQL supports outer references in derived tables,
if the derived table is merged we end up with a column of type
Item_view_ref with used_tables() having OUTER_REF_TABLE_BIT.
In a place of the functional dependency logic which validates
grouped queries, we didn't expect this bit.
Fix: just clear this bit; in this place we're interested
in knowing local (non-outer) tables referenced by the column, only.

Reviewed by: Sergey Glukhov <sergey.glukhov@oracle.com>

Change-Id: If1f181cf8a5f38855f0d9d213ee26bd3c2e96675