Bug #30368937: MATERIALIZED SUBQUERY PRODUCES WRONG RESULT WITH A FUNCTION
Description:
A query with a non-deterministic funtion in it's subquery produces
wrong result when the subquery is materialized.

ANALYSIS:
WHERE condition of materialized subquery containing
non-deterministic function is not being attached to the table. This
is due to the condition "cond->used_tables() & ~tables" in
make_cond_for_table_from_pred(). Where tables (tab->prefix_tables())
is not being set properly through JOIN::set_prefix_tables().

Purpose of JOIN::set_prefix_tables() is to add non-deterministic
functions to the "last table" in a join nest by adding
RAND_TABLE_BIT. However with materialized subqueries, we have a
"last table" both in the main query as well as in each materialized
subquery. But set_prefix_tables() fails to add this bit for
subqueries.

SOLUTION:
Modify JOIN::set_prefix_tables() to add RAND_TABLE_BIT to the last
table of each materialized subquery.

Change-Id: Ie31c62264f4dfa08fc1250b3d6cc2f2b4cbc124c