Bug #30791074: WEIGHT_STRING: ASSERTION `0' FAILED. IN TIME_ZONE_UTC::TIME_TO_GMT_SEC
In bug #30776132 (MAKE FILESORT KEYS CONSISTENT BETWEEN FIELDS AND ITEMS),
we made filesort unconditionally switch to UTC when evaluating timestamps.
However, this was made by setting the time zone to a global of type
Time_zone_utc. This class is not the same as the regular UTC time zone
(typically of type Time_zone_system or similar), and was missing one important
conversion, namely that of MYSQL_TIME to my_time_t, causing wrong results
and also assertion failures in debug mode. The conversion is trivial,
so we simply add it.

Change-Id: I807f189973a3a22fdf6040b95c0862b2c68e97f9