WL#13174: CREATE NEW RUNS OF EXISTING TESTS WITH RESTRICTED PRIVILEGES FOR REPLICATION
Make it possible to run replication tests with restricted privileges for the
replication applier so that regressions introduced by the design and
implementation of privilege checking can be detected.

WL12966 "Replication with privilege checks" made it possible to enable privilege
checks on replication channels, so that it is secure to replicate from untrusted
masters. Database changes originating from the master can be applied in the
slave using a privilege context.

Isolated the configuration and setup of replication applier privilege context in
specific and re-usable include files, which can be used by any newly added
test-scripts.

To run a test suite with privilege checks for the replication applier, execute:
./mysql-test-run.pl --suite=<name>
--defaults-extra-file=include/rpl_priv_checks_user_no_dynamic_grants.cnf

RB: 23147,23419,23445,23446,23483,23482,23695
Reviewed-by: Pedro Figueiredo <pedro.figueiredo@oracle.com>
Reviewed-by: Nuno Carvalho <nuno.carvalho@oracle.com>
Reviewed-by: Parveez Baig <parveez.baig@oracle.com>
Reviewed-by: Pedro Gomes <pedro.gomes@oracle.com>