BUG#30259491 IMPROVE CLUSTER PROCESSES MANAGEMENT
Change when and how cluster processes are started and stopped to improve
handling of critical error when running testcases. Provide a new high
level API in class ProcessManagement for managing processes.

The introduced changes ensure that the cluster processes are always
started, if needed, at the beginning of the execution of a testcase (or
a retry in case of failure). Moreover, all cluster processes are stopped
at the end of the execution of a testcase in the following conditions:
- it is the last testcase;
- the next testcase forces a cluster restart by configuration;
- the configuration has changed (upgrade/downgrade tests);
- the test has failed and the configure behavior is to stop the cluster
processes.
Otherwise, only the client processes will be stopped. All this will
ensure that changes in the cluster status are confined to the context of
the execution of a testcase, and all cluster logs are stored in the case
of a failure during start/stop of processes.

To better handle this approach, a variable to track the status of the
cluster processes was added to the ProcessManagement class. This way,
when the global status is RUNNING, calls to start cluster processes
succeed immediately without performing any action. The same thing
happens when the status is STOPPED and a call to stop cluster processes
is done. When stopping all processes or client processes fails, the
global status is set to ERROR since it is neither running or stopped.
This will ensure that next time a call to start cluster processes is
made, there will be an attempt to stop all processes. If this fails, the
start operation will also fail, not allowing to run any tests with an
inconsistent state.

Additionally, new high level methods in class ProcessManagement were
provided making a more coherent API. Most of the existing methods were
made private and left unchanged, although they would benefit of a
refactoring in the future. Methods updateStatus and
checkNdbOrServersFailures were merged since they are always used
together.

The implementation of high level methods of ProcessManagement followed
the principle that logging of errors should be done by the caller. In
methods composed by more than one call to other methods, a warning is
logged to help determining the failing point when debugging.

Function run_test_case was modified to accept two boolean parameters to
signal two of the conditions required to stop the cluster processes at
the end of execution: if the current test is the last one or if the next
testcase forces a restart of the cluster by configuration. For this, a
single boolean parameter would be enough but having both cases
explicitly stated makes reading the cases for stopping the cluster
processes clearer. Similarly, when computing these values for calling
the run_test_cases function, it would be enough to use the expression
'last_testcase_idx == i || testcases[i+1].m_force_cluster_restart'.
However, this reduces the understandability of why it is being used.

Change-Id: I4fd94a5415afd1ccec07ab992343c2797e0a8097