Bug#30717778 WL13620: ASSERTION FAILURE: RET == 0 IN OS0EVENT.CC ON REFRESHING PRELOAD STATS - Issue 1
Issue
=====
The histogram sampler makes use of the parallel read infrastructure and as part
of this infrastructure we spawn a single thread as we do not support parallel
read yet for sampling. This thread acts as producer where it reads all the
records which the main thread consumes via a buffer.

In case of errors the threads are synchronised using a central error state.
Either the reader thread communicates to the main thread that there is an error
or the main thread communicates to the reader thread to abort the operation.
The issue is that we seem to set this error state even in the case of success,
during the operation of the reader thread, and this can result in a deadlock
situation where the thread has been asked to abort but it doesn't abort because
the error state has been reset to success.

Fix
===
The fix is to set the error state only in case of an error.
mysql-trunk-wl8777-ap

RB: 23708
Reviewed-by: Annamalai Gurusami <annamalai.gurusami@oracle.com>