Bug#30541701: WRONG RESULT WITH ROLLUP Bug#30761372: THE "ORDER BY" OPTION CAUSE WRONG RESULT
Bug#30541701
Problem:
ROLLUP fails to output nulls when expressions in group by have column names
different in case from the actual columns created in the table
(For ex: group by has "id" and the actual column in table is "ID").

Analysis:
While resolving group by expressions, when the expressions are not found in
the select list, a hidden item is added to the fields list needed by the
query. For the case in bug page, even though its the same expression found
in select list and group by, a hidden item is created because the actual
columns created in the table have different case like "iD and INT_TEST"
where as the group by expression has "id and int_test". As a result,
find_item_in_list fails to find the expression in the select list(because
by then this expression is already resolved against the column names
in the table) and therefore a hidden item gets added to the fields list
of the query.
Later, while setting up copy fields,  hidden item and the item in select
list get replaced with different "Item_copy_*" objects.
These items are later compared in rollup_make_fields in determining
if a NULL needs to be replaced in the rollup output row. Each of the matching
expression (to the one in group by) gets replaced with a "Item_null_result"
object in the rollup "ref_item_array". However since the grouped expression
and the one in the select list do not have same "Item_copy_*" objects, the
comparison fails resulting in no "NULL" in the output.

Solution:
For expressions whose column names only differ in "case", Item_func::eq
returns false as it does a "case sensitive" comparison. However, it is
expected to do a "case insensitive" comparison because columns in expressions
are resolved against columns in tables using "case insensitive" comparison.
This inconsistency is what results in creation of hidden items which are not
only not necessary, but also problematic because once the fields are resolved,
checks for equality between these fields always returns true. This leads to
the use of hidden item sometimes and the item in the select list sometimes.

Bug#30761372:

Problem is similar to what is stated above. For a query, which has distinct,
group by and order by, if an expression in select list has a column with
its name differing only in case from the column in the table, then distinct
removes more rows.

Analysis:
Similar to the rollup issue above, a hidden item is added initially because
of the "case sensitive" comparison. This leads to inconsistencies later
(create_order_from_distinct()) as the order by expression would be marked
as not present in the select list. However when the expressions in select
list are compared with expressions in order by, it succeeds. This results
in a missing expression for duplicate removal leading to missing rows from
the output.

Reviewed By : Dag Wanvik <dag.wanvik@oracle.com>