BUG#30805827: BACKPORT BUG#30799916 TO 5.7
rpl_transaction_write_set_extraction test was introduced to validate
that the write-set string value computed by
transaction_write_set_extraction option is deterministic
independently of the server version or platform on which it is
generated.

This patch extends rpl_transaction_write_set_extraction test to also
validate that the hash value generated by XXHASH64 algorithm is
deterministic, independently of the server version or platform.

RB: 23731