Bug#30717778 WL13620: ASSERTION FAILURE: RET == 0 IN OS0EVENT.CC ON REFRESHING PRELOAD STATS - Issue 2
Issue
=====
In the histogram sampling code there is a case where we prematurely exit
sampling (feature requested by RAPID) and this scenario is handled by toggling
a boolean variable Histogram_sampler::m_sampling_done. The issue is that this
is set by both the main thread and the reader worker thread which results in a
stall where the main thread thinks the reader thread is done with its work and
has exited and tries to destroy the object associated with it.

Fix
===
On closer look it was seen that Histogram_sampler::m_sampling_done was not
really required.

The usual behavior is that the worker thread waits for a buffer event request
from the main thread and when the main thread wants a record it signals this
event. There are two special cases which needs to be handled.

a) Error case where the worker thread errors out and exits and the main thread
is not aware of it before it requests for another record.
b) Case where the server decides to prematurely exit sampling and calls the
cleanup function (sample_end)

a) is fixed by having the thread wait for the event in the callback that the
thread calls when it's exiting. The main thread would signal the worker thread
and worker thread would set a central error state and exit. The main thread
would then check for this error state and exit.
b) is handled by the sample_end function signaling the worker thread to exit
and waiting for it to exit.

Code for both a) and b) are already present in the codebase and there was no
need of Histogram_sampler::m_sampling_done. So this patch removes it
altogether.

RB: 23727
Reviewed-by: Annamalai Gurusami <annamalai.gurusami@oracle.com>