Bug#29952565 where MYSQLD GOT SIGNAL 11 WHILE EXECUTING A QUERY(UNION + ORDER BY + SUB-QUERY)
Problem:
Removal of a subquery which is part of query's select list (projection)
and also order by results in server exit.

For a query, order by can be considered redundant if it is part of a
union statement. order by is therefore removed and in doing so, the
subquery is also removed. Since the subquery is part of select list, it
can be accessed later and therefore should not be removed.

Fix:
Do not remove the subquery if it is part of the select list.

Change-Id: If1b3c96f4749e9313e3579beeab49bd7e60148fb