Bug#30158954: INCONSISTENT ROWS MATCHED VALUE FOR UPDATE WITH FAILED CONSTRAINT
The problem was that UPDATE statements could give an inconsistent
number of rows matched (found rows) in cases where not all matched
rows were updated depending on the reason for rows not being updated.

For example, rows that where not updated due to being updated
through a view with a WITH CHECK OPTION clause, are not counted
as matching rows. However, rows that were not updated due to a
failing CHECK CONSTRAINT were counted as matching rows.

This patch makes the behavior more consistent by counting rows
that fail a WITH CHECK OPTION clause as "matching rows".

Change-Id: I316965445dcd687fe0bd4ee6fb9fddd17ddb9fce