Bug#30814171: VALGRIND LEAK FOUND AT STRING::REAL_ALLOC() IN SQL_STRING.CC
The problem was that queries with IN expressions that acccessed
columns with large string values could lead to a memory leak.

The root cause was that the destructor for in_row::tmp was not
called, which could lead to the destructor for cmp_item_string::value
not being called. This patch fixes the problem by making sure
the relevant destructors are called.

This was a regression introduced by the patch for Bug#28661915.

Change-Id: I26d76410dd8d7b1ffdf1027506f00a1bcc0a5c35