WL13065: Restrict All InnoDB Tablespaces To Known Directories
1. Make CREATE TABLE ... DATA DIRECTORY to fail if the directory specified
is not a known IBD location. This restriction is already done for
CREATE TABLESPACE ... ADD DATAFILE.

2. Startup does not have any new restrictions to where files can be
found, but there are new error log messages that complain about any
file referenced in the DD but that are not in known directories.

2a. TRUNCATE TABLE on a table in an unknown location will create
the new file in the default datadir with an appropriate warning
message in the log.

3. If innodb-undo-directory is a unique location, it will not be
considered a known IBD location for the purpose of creating new
datafiles.

4. Many testcases that use DATA DIRECTORY have been modified to add
that location to --innodb-directories.

Approved in RB#22303 by Annamalai and Mayank